[#124] Move parsing of data to reporting package
